pull out AddFishForm and Inventory
